+++ CHANGELOG.md
@@ -7,7 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-No changes yet.
+- Add mouse navigation buttons and a keyboard-accessible destination popup to the title bar.
+- Allow Global Members and Commit Search rows to select and open with repeated clicks.
 
 ## [0.4.0] - 2026-09-05